writer and bit width updates

+++ src/runtime_src/xdp/profile/writer/aie_debug/aie_debug_writer_metadata.h
@@ -36,9 +36,9 @@ class WriterUsedRegisters {
       std::string field_name;
       std::string bit_range;
       int shift;
-      uint64_t mask;
+      uint32_t mask;
 
-      RegData(std::string n, std::string b, int s, uint64_t m)
+      RegData(std::string n, std::string b, int s, uint32_t m)
         : field_name(n), bit_range(b), shift(s), mask(m) {}
     };
