+++ src/runtime_src/xdp/profile/writer/aie_debug/register_interpreter.h
@@ -32,13 +32,13 @@ class RegisterInterpreter {
     struct RegInfo {
       std::string field_name;
       std::string bit_range;
-      uint64_t subval;
+      uint32_t subval;
 
-      RegInfo(std::string f, std::string b, uint64_t s)
+      RegInfo(std::string f, std::string b, uint32_t s)
         : field_name(f), bit_range(b), subval(s) {}
     };
 
-    std::vector<RegInfo> registerInfo(const std::string &regName, const uint64_t &regAddr, const uint64_t &regVal);
+    std::vector<RegInfo> registerInfo(const std::string &regName, const uint64_t &regAddr, const uint32_t &regVal);
 
   private:
     std::unique_ptr<WriterUsedRegisters> writerUsedRegisters;
